refactor(swift): extract compileProtobufs function to its own file
